Always free the MDB_Pages
It was an issue to keep the pages in a chained cache list,
it was growing a lot and consumming a lot of memory.

The only way to free this chained list was to close the environment,
this solution was a little bit too much, so we now entirely skip the cache
and let malloc to that for us, by itself.