Allow multiple read-only txns from a transaction
We accept nested read-only transactions and no longer count them to
accept more than a single one. We make sure that we free/close the
child transactions only once the last one has been dropped.

We use atomics to implement a reference-counted system to correctly
drop the read transactions and the pghead only once. This is not C99
but C11-compatible.